requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9
Pillow>=10.0
numpy>=1.24
scikit-image>=0.22
pillow-avif-plugin>=1.4
